problem: (nonfunctional,refactor) should move cmd-oriented functions to cmd.go
solution: cut and paste and update imports
