+++ cmd/geth/cmd.go
@@ -34,6 +34,15 @@ import (
 	"github.com/ethereumproject/go-ethereum/rlp"
 	"time"
 	"syscall"
+	"math/big"
+	"gopkg.in/urfave/cli.v1"
+	"strconv"
+	"github.com/ethereumproject/go-ethereum/eth/downloader"
+	"path/filepath"
+	"io/ioutil"
+	"github.com/ethereumproject/ethash"
+	"github.com/ethereumproject/go-ethereum/core/state"
+	"errors"
 )
 
 const (
@@ -482,4 +491,448 @@ func formatChainDataPretty(datadir string, chain *core.BlockChain) (s []string) 
 		}
 	}
 	return s
+}
+
+func status(ctx *cli.Context) error {
+
+	shouldUseExisting := false
+	datadir := MustMakeChainDataDir(ctx)
+	chaindatadir := filepath.Join(datadir, "chaindata")
+	if di, e := os.Stat(chaindatadir); e == nil && di.IsDir() {
+		shouldUseExisting = true
+	}
+	// Makes sufficient configuration from JSON file or DB pending flags.
+	// Delegates flag usage.
+	config := mustMakeSufficientChainConfig(ctx)
+
+	// Configure the Ethereum service
+	ethConf := mustMakeEthConf(ctx, config)
+
+	// Configure node's service container.
+	name := makeNodeName(Version, ctx)
+	stackConf, _ := mustMakeStackConf(ctx, name, config)
+
+	sep := glog.Separator("-")
+	printme := []struct {
+		title   string
+		keyVals []string
+	}{
+		{"Chain configuration", formatSufficientChainConfigPretty(config)},
+		{"Ethereum configuration", formatEthConfigPretty(ethConf)},
+		{"Node configuration", formatStackConfigPretty(stackConf)},
+	}
+
+	s := "\n"
+
+	for _, p := range printme {
+		s += withLineBreak(sep)
+		// right align category title
+		s += withLineBreak(strings.Repeat(" ", len(sep)-len(p.title)) + colorBlue(p.title))
+		for _, v := range p.keyVals {
+			s += v
+		}
+	}
+	glog.V(logger.Info).Info(s)
+
+	// Return here if database has not been initialized.
+	if !shouldUseExisting {
+		glog.V(logger.Info).Info("Geth has not been initialized; no database information available yet.")
+		return nil
+	}
+
+	chaindata, cdb := MakeChain(ctx)
+	defer cdb.Close()
+	s = "\n"
+	s += withLineBreak(sep)
+	title := "Chain database status"
+	s += withLineBreak(strings.Repeat(" ", len(sep)-len(title)) + colorBlue(title))
+	for _, v := range formatChainDataPretty(datadir, chaindata) {
+		s += v
+	}
+	glog.V(logger.Info).Info(s)
+
+	return nil
+}
+
+func rollback(ctx *cli.Context) error {
+	index := ctx.Args().First()
+	if len(index) == 0 {
+		glog.Fatal("missing argument: use `rollback 12345` to specify required block number to roll back to")
+		return errors.New("invalid flag usage")
+	}
+
+	blockIndex, err := strconv.ParseUint(index, 10, 64)
+	if err != nil {
+		glog.Fatalf("invalid argument: use `rollback 12345`, were '12345' is a required number specifying which block number to roll back to")
+		return errors.New("invalid flag usage")
+	}
+
+	bc, chainDB := MakeChain(ctx)
+	defer chainDB.Close()
+
+	glog.Warning("Rolling back blockchain...")
+
+	bc.SetHead(blockIndex)
+
+	nowCurrentState := bc.CurrentBlock().Number().Uint64()
+	if nowCurrentState != blockIndex {
+		glog.Fatalf("ERROR: Wanted rollback to set head to: %v, instead current head is: %v", blockIndex, nowCurrentState)
+	} else {
+		glog.Infof("SUCCESS: Head block set to: %v", nowCurrentState)
+	}
+
+	return nil
+}
+
+// dumpChainConfig exports chain configuration based on context to JSON file.
+// It is not compatible with --chain-config flag; it is intended to move from flags -> file,
+// and not the other way around.
+func dumpChainConfig(ctx *cli.Context) error {
+
+	chainIdentity := mustMakeChainIdentity(ctx)
+	if !(chainIdentitiesMain[chainIdentity] || chainIdentitiesMorden[chainIdentity]) {
+		glog.Fatal("Dump config should only be used with default chain configurations (mainnet or morden).")
+	}
+
+	glog.V(logger.Info).Infof("Dumping configuration for: %v", chainIdentity)
+
+	chainConfigFilePath := ctx.Args().First()
+	chainConfigFilePath = filepath.Clean(chainConfigFilePath)
+
+	if chainConfigFilePath == "" || chainConfigFilePath == "/" || chainConfigFilePath == "." {
+		glog.Fatalf("Given filepath to export chain configuration was blank or invalid; it was: '%v'. It cannot be blank. You typed: %v ", chainConfigFilePath, ctx.Args().First())
+		return errors.New("invalid required filepath argument")
+	}
+
+	fb := filepath.Dir(chainConfigFilePath)
+	di, de := os.Stat(fb)
+	if de != nil {
+		if os.IsNotExist(de) {
+			glog.V(logger.Warn).Infof("Directory path '%v' does not yet exist. Will create.", fb)
+			if e := os.MkdirAll(fb, os.ModePerm); e != nil {
+				glog.Fatalf("Could not create necessary directories: %v", e)
+			}
+			di, _ = os.Stat(fb) // update var with new dir info
+		} else {
+			glog.V(logger.Error).Infof("err: %v (at '%v')", de, fb)
+		}
+	}
+	if !di.IsDir() {
+		glog.Fatalf("'%v' must be a directory", fb)
+	}
+
+	// Implicitly favor Morden because it is a smaller, simpler configuration,
+	// so I expect it to be used more frequently than mainnet.
+	genesisDump := core.TestNetGenesis
+	netId := 2
+	stateConf := &core.StateConfig{StartingNonce: state.DefaultTestnetStartingNonce}
+	if !chainIsMorden(ctx) {
+		genesisDump = core.DefaultGenesis
+		netId = eth.NetworkId
+		stateConf = nil
+	}
+
+	// Note that we use default configs (not externalizable).
+	chainConfig := MustMakeChainConfigFromDefaults(ctx)
+	var nodes []string
+	for _, node := range MakeBootstrapNodesFromContext(ctx) {
+		nodes = append(nodes, node.String())
+	}
+
+	var currentConfig = &core.SufficientChainConfig{
+		Identity:    chainIdentity,
+		Name:        mustMakeChainConfigNameDefaulty(ctx),
+		Network:     netId,
+		State:       stateConf,
+		Consensus:   "ethash",
+		Genesis:     genesisDump,
+		ChainConfig: chainConfig.SortForks(), // get current/contextualized chain config
+		Bootstrap:   nodes,
+	}
+
+	if writeError := currentConfig.WriteToJSONFile(chainConfigFilePath); writeError != nil {
+		glog.Fatalf("An error occurred while writing chain configuration: %v", writeError)
+		return writeError
+	}
+
+	glog.V(logger.Info).Info(fmt.Sprintf("Wrote chain config file to \x1b[32m%s\x1b[39m.", chainConfigFilePath))
+	return nil
+}
+
+// startNode boots up the system node and all registered protocols, after which
+// it unlocks any requested accounts, and starts the RPC/IPC interfaces and the
+// miner.
+func startNode(ctx *cli.Context, stack *node.Node) *eth.Ethereum {
+	// Start up the node itself
+	StartNode(stack)
+
+	// Unlock any account specifically requested
+	var ethereum *eth.Ethereum
+	if err := stack.Service(&ethereum); err != nil {
+		glog.Fatal("ethereum service not running: ", err)
+	}
+
+	// Start auxiliary services if enabled
+	if ctx.GlobalBool(aliasableName(MiningEnabledFlag.Name, ctx)) {
+		if err := ethereum.StartMining(ctx.GlobalInt(aliasableName(MinerThreadsFlag.Name, ctx)), ctx.GlobalString(aliasableName(MiningGPUFlag.Name, ctx))); err != nil {
+			glog.Fatalf("Failed to start mining: %v", err)
+		}
+	}
+	return ethereum
+}
+
+func makedag(ctx *cli.Context) error {
+	args := ctx.Args()
+	wrongArgs := func() {
+		glog.Fatal(`Usage: geth makedag <block number> <outputdir>`)
+	}
+	switch {
+	case len(args) == 2:
+		blockNum, err := strconv.ParseUint(args[0], 0, 64)
+		dir := args[1]
+		if err != nil {
+			wrongArgs()
+		} else {
+			dir = filepath.Clean(dir)
+			// seems to require a trailing slash
+			if !strings.HasSuffix(dir, "/") {
+				dir = dir + "/"
+			}
+			_, err = ioutil.ReadDir(dir)
+			if err != nil {
+				glog.Fatal("Can't find dir")
+			}
+			fmt.Println("making DAG, this could take awhile...")
+			ethash.MakeDAG(blockNum, dir)
+		}
+	default:
+		wrongArgs()
+	}
+	return nil
+}
+
+func gpuinfo(ctx *cli.Context) error {
+	eth.PrintOpenCLDevices()
+	return nil
+}
+
+func gpubench(ctx *cli.Context) error {
+	args := ctx.Args()
+	wrongArgs := func() {
+		glog.Fatal(`Usage: geth gpubench <gpu number>`)
+	}
+	switch {
+	case len(args) == 1:
+		n, err := strconv.ParseUint(args[0], 0, 64)
+		if err != nil {
+			wrongArgs()
+		}
+		eth.GPUBench(n)
+	case len(args) == 0:
+		eth.GPUBench(0)
+	default:
+		wrongArgs()
+	}
+	return nil
+}
+
+func version(ctx *cli.Context) error {
+	fmt.Println("Geth")
+	fmt.Println("Version:", Version)
+	fmt.Println("Protocol Versions:", eth.ProtocolVersions)
+	fmt.Println("Network Id:", ctx.GlobalInt(aliasableName(NetworkIdFlag.Name, ctx)))
+	fmt.Println("Go Version:", runtime.Version())
+	fmt.Println("OS:", runtime.GOOS)
+	fmt.Printf("GOPATH=%s\n", os.Getenv("GOPATH"))
+	fmt.Printf("GOROOT=%s\n", runtime.GOROOT())
+
+	return nil
+}
+
+// LogStatusFeatAvailability is used to register and track use of available status logging features, eg. "STATUS SYNC"
+type LogStatusFeatAvailability int
+const (
+	StatusFeatAvailable   LogStatusFeatAvailability = iota
+	StatusFeatRegistered
+	StatusFeatNonexistent
+)
+
+// availableLogStatusFeatures stores state of implemented log STATUS features.
+// New features should be registered here, and their status updates by dispatchStatusLogs if in use (to avoid dupe goroutine logging).
+var availableLogStatusFeatures = map[string]LogStatusFeatAvailability{
+	"sync": StatusFeatAvailable,
+}
+
+// dispatchStatusLogs handle parsing --log-status=argument and toggling appropriate goroutine status feature logging.
+func dispatchStatusLogs(ctx *cli.Context, ethe *eth.Ethereum) {
+	flagName := aliasableName(LogStatusFlag.Name, ctx)
+	v := ctx.GlobalString(flagName)
+	if v == "" {
+		glog.Fatalf("%v: %v", flagName, ErrInvalidFlag)
+	}
+
+	for _, p := range strings.Split(v, ",") {
+		// Ignore hanging or double commas
+		if p == "" {
+			continue
+		}
+
+		// If possible, split sync=60 into ["sync", "60"], otherwise yields ["sync"], ["60"], or ["someothernonsense"]
+		eqs := strings.Split(p, "=")
+
+		// Catch unavailable and duplicate status feature logs
+		if availableLogStatusFeatures[eqs[0]] == StatusFeatNonexistent {
+			glog.Fatalf("%v: %v: unavailable status feature by name of '%v'", flagName, ErrInvalidFlag, eqs[0])
+		}
+		if availableLogStatusFeatures[eqs[0]] == StatusFeatRegistered {
+			glog.Fatalf("%v: %v: duplicate status feature by name of '%v'", flagName, ErrInvalidFlag, eqs[0])
+		}
+
+		// If user just uses "sync" instead of "sync=42", append empty string and delegate to each status log function how to handle it
+		if len(eqs) == 1 {
+			eqs = append(eqs, "")
+		}
+		switch eqs[0] {
+		case "sync":
+			availableLogStatusFeatures["sync"] = StatusFeatRegistered
+			go runStatusSyncLogs(ethe, eqs[1], ctx.GlobalInt(aliasableName(MaxPeersFlag.Name, ctx)))
+		}
+	}
+}
+
+// runStatusSyncLogs starts STATUS SYNC logging at a given interval.
+// It should be run as a goroutine.
+// eg. --log-status="sync=42" logs SYNC information every 42 seconds
+func runStatusSyncLogs(e *eth.Ethereum, interval string, maxPeers int) {
+
+	// Establish default interval.
+	intervalI := 60
+
+	if interval != "" {
+		i, e := strconv.Atoi(interval)
+		if e != nil {
+			glog.Fatalf("STATUS SYNC %v: could not parse argument: %v", e, interval)
+		}
+		if i < 1 {
+			glog.Fatalf("STATUS SYNC interval value must be a positive integer, got: %d", i)
+		}
+		intervalI = i
+	}
+
+	glog.V(logger.Error).Infof("STATUS SYNC Log-pace interval set: %d seconds", intervalI)
+
+	tickerInterval := time.Second * time.Duration(int32(intervalI))
+	ticker := time.NewTicker(tickerInterval)
+
+	var lastLoggedBlockNumber uint64
+
+	for {
+		select {
+		case <-ticker.C:
+			lenPeers := e.Downloader().GetPeers().Len()
+
+			_, current, height, _, _ := e.Downloader().Progress() // origin, current, height, pulled, known
+			mode := e.Downloader().GetMode()
+
+			// Get our head block
+			blockchain := e.BlockChain()
+			currentBlockHex := blockchain.CurrentBlock().Hash().Hex()
+
+			// Discover -> not synchronising (searching for peers)
+			// FullSync/FastSync -> synchronising
+			// Import -> synchronising, at full height
+			fMode := "Discover"
+			fOfHeight := fmt.Sprintf(" of #%7d", height)
+
+			// Calculate and format percent sync of known height
+			heightRatio := float64(current) / float64(height)
+			heightRatio = heightRatio * 100
+			fHeightRatio := fmt.Sprintf("(%4.2f%%)", heightRatio)
+
+			// Wait until syncing because real dl mode will not be engaged until then
+			if e.Downloader().Synchronising() {
+				switch mode {
+				case downloader.FullSync:
+					fMode = "FullSync"
+				case downloader.FastSync:
+					fMode = "FastSync"
+					currentBlockHex = blockchain.CurrentFastBlock().Hash().Hex()
+				}
+			}
+			if current == height && !(current == 0 && height == 0) {
+				fMode = "Import  " // with spaces to make same length as Discover, FastSync, FullSync
+				fOfHeight = strings.Repeat(" ", 12)
+				fHeightRatio = strings.Repeat(" ", 7)
+			}
+			if height == 0 {
+				fOfHeight = strings.Repeat(" ", 12)
+				fHeightRatio = strings.Repeat(" ", 7)
+			}
+
+			// Calculate block stats for interval
+			numBlocksDiff := current - lastLoggedBlockNumber
+			numTxsDiff := 0
+			mGas := new(big.Int)
+
+			var numBlocksDiffPerSecond uint64
+			var numTxsDiffPerSecond int
+			var mGasPerSecond = new(big.Int)
+
+			if numBlocksDiff > 0 && numBlocksDiff != current {
+				for i := lastLoggedBlockNumber; i <= current; i++ {
+					b := blockchain.GetBlockByNumber(i)
+					if b != nil {
+						numTxsDiff += b.Transactions().Len()
+						mGas = new(big.Int).Add(mGas, b.GasUsed())
+					}
+				}
+			}
+
+			// Convert to per-second stats
+			// FIXME(?): Some degree of rounding will happen.
+			// For example, if interval is 10s and we get 6 blocks imported in that span,
+			// stats will show '0' blocks/second. Looks a little strange; but on the other hand,
+			// precision costs visual space, and normally just looks weird when starting up sync or
+			// syncing slowly.
+			numBlocksDiffPerSecond = numBlocksDiff / uint64(intervalI)
+
+			// Don't show initial current / per second val
+			if lastLoggedBlockNumber == 0 {
+				numBlocksDiffPerSecond = 0
+			}
+
+			// Divide by interval to yield per-second stats
+			numTxsDiffPerSecond = numTxsDiff / intervalI
+			mGasPerSecond = new(big.Int).Div(mGas, big.NewInt(int64(intervalI)))
+			mGasPerSecond = new(big.Int).Div(mGasPerSecond, big.NewInt(1000000))
+			mGasPerSecondI := mGasPerSecond.Int64()
+
+			// Update last logged current block number
+			lastLoggedBlockNumber = current
+
+			// Format head block hex for printing (eg. d4e…fa3)
+			cbhexstart := currentBlockHex[2:5] // trim off '0x' prefix
+			cbhexend := currentBlockHex[(len(currentBlockHex) - 3):]
+
+			blockprogress := fmt.Sprintf("#%7d%s", current, fOfHeight)
+			cbhexdisplay := fmt.Sprintf("%s…%s", cbhexstart, cbhexend)
+			peersdisplay := fmt.Sprintf("%2d/%2d peers", lenPeers, maxPeers)
+			blocksprocesseddisplay := fmt.Sprintf("%4d/%4d/%2d blks/txs/mgas sec", numBlocksDiffPerSecond, numTxsDiffPerSecond, mGasPerSecondI)
+
+			// Log to ERROR.
+			// This allows maximum user optionality for desired integration with rest of event-based logging.
+			glog.V(logger.Error).Infof("STATUS SYNC %s %s %s %s %s %s", fMode, blockprogress, fHeightRatio, cbhexdisplay, blocksprocesseddisplay, peersdisplay)
+
+			// Experimental: if mining, show some readily-available stats for that, too
+			miner := e.Miner()
+			if miner.Mining() {
+				hashrate := miner.HashRate()
+				glog.V(logger.Error).Infof("STATUS MINE (%2d) %8d", e.MinerThreads, hashrate)
+			}
+			// Listen for interrupt
+		case <-sigc:
+			ticker.Stop()
+			return
+		}
+	}
 }
